+++ backend/src/main/java/mouda/backend/notification/implement/RecipientFactory.java
@@ -1,4 +1,4 @@
-package mouda.backend.notification.business;
+package mouda.backend.notification.implement;
 
 import java.util.List;
 import java.util.Map;
@@ -8,14 +8,13 @@ import org.springframework.stereotype.Service;
 
 import mouda.backend.notification.domain.NotificationType;
 import mouda.backend.notification.domain.NotificationTypeProvider;
-import mouda.backend.notification.domain.notification.NotificationBuilderStrategy;
+import mouda.backend.notification.domain.recipient.RecipientResolverStrategy;
 
 @Service
-public class NotificationFactory {
+public class RecipientFactory {
+	private final Map<NotificationType, RecipientResolverStrategy> strategies;
 
-	private final Map<NotificationType, NotificationBuilderStrategy> strategies;
-
-	public NotificationFactory(List<NotificationBuilderStrategy> strategyList) {
+	public RecipientFactory(List<RecipientResolverStrategy> strategyList) {
 		this.strategies = strategyList.stream()
 			.collect(
 				Collectors.toMap(strategy -> {
@@ -24,7 +23,7 @@ public class NotificationFactory {
 					strategy -> strategy));
 	}
 
-	public NotificationBuilderStrategy getStrategy(NotificationType type) {
+	public RecipientResolverStrategy getStrategy(NotificationType type) {
 		return strategies.get(type);
 	}
 }
